v7.2.0
- Bring date's methods back
- Update test scripts
- Enable test with node 9
