+++ .travis.yml
@@ -1,6 +1,7 @@
 language: node_js
 node_js:
-  - "8.2"
+  - "9.0"
+  - "8.0"
   - "7.0"
   - "6.0"
 before_script:
