+++ package.json
@@ -1,5 +1,5 @@
 {
-  "version": "7.1.3",
+  "version": "7.2.0",
   "name": "bellajs",
   "description": "A useful helper for any javascript program",
   "homepage": "https://www.npmjs.com/package/bellajs",
@@ -39,6 +39,7 @@
     "jsdom-global": "3.x.x",
     "nsp": "2.x.x",
     "nyc": "11.x.x",
+    "sinon": "4.x.x",
     "tap-spec": "4.x.x",
     "tape": "4.x.x"
   },
